DeviceMgr: Platform Invoke should use correct CharSet
The CharSet should be CharSet.Unicode, instead of CharSet.Auto. The API P/Invokes that weren't used were removed.

+++ DeviceMgr/Native/Win32/Kernel32+GetOsVersion.cs
@@ -28,7 +28,7 @@
                     return Environment.OSVersion.Version;
                 }
 
-                if (info.PlatformId != (int)WinPlatformId.WinNT) {
+                if (info.PlatformId != WinPlatformId.WinNT) {
                     throw new PlatformNotSupportedException();
                 }
